Basically finished
The basic game is working now. Arrow keys control the snake. Collecting the blue blocks increases the size of the snake. Hitting a wall or yourself will kill the snake, turning its head red. Score is based on the number of blocks collected.

If im still bored I might work on other features like text rendering for the score so I can get rid of the console window. Also speed increase as the snake size increases might be interesting.