minor bugfix
Was just bored and decided to play some snake. Got pretty far too until I encountered an exception. Game over man. Anyway it turned out that a new objective was created with the same grid index as another objective. This new objective was duplicated in Snake.objective_list but not in Grid.active_grid_elements. We can't have two keys in a dictionary that are the same so it ended up overwriting the old data in active_grid_elements. The objective was eventually collected, deleting its entry in active_grid_elements but only one of the duplicates in objective_list was removed. When it came time to remove old objectives, this left-over objective grid was pop'd to be removed. Unfortunately that grid no longer existed in active_grid_elements which caused an Exception. And now you know the rest of the story....