Removed defunct code
Accidentally left some code in there from the previous app I pulled some bits from. In the old app I had to calculate the vertices exactly in a reference frame that was possibly not axis-aligned. This meant it was mathematically easier to generate the vertices by knowing the absolute center and the current rotation of the object. Then it was very simple maths to figure out each vertex. This code in full is in one of my public gists for BouncySquare. Here we know each grid element is axis-aligned so there is no need to calculate the absolute center.